requirements.txt
six
numpy>=1.16.0,<2.0
torch>=1.8.0,<1.9.0
torchvision==0.9.0
tqdm
pillow
PyYAML
flask
boto3
pytz